# webui/openapi.yaml
openapi: 3.1.0
info:
  title: KToolBox WebUI API
  description: Authenticated REST API for the KToolBox WebUI. Selected operations also define the curated MCP tool surface
    through x-ktoolbox-mcp metadata.
  version: 1.0.0
  license:
    name: BSD-3-Clause
    identifier: BSD-3-Clause
paths:
  /api/v1/config/schema:
    get:
      summary: Get configuration schema
      operationId: config_schema
      parameters:
      - name: locale
        in: query
        required: false
        schema:
          enum:
          - zh-CN
          - zh-Hant
          - en
          - ja
          - ko
          - fr
          - ru
          type: string
          default: en
          title: Locale
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigSchemaResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      description: Return localized field metadata and redacted effective values for the project configuration.
      tags:
      - configuration
      security:
      - WebUISession: []
      x-ktoolbox-mcp:
        enabled: true
        name: config_schema
        scope: mcp:read
        safety: read
        openWorld: false
...
  /api/v1/openapi.yaml:
    get:
      tags:
      - mcp
      summary: Download WebUI OpenAPI
      operationId: download_openapi
      responses:
        '200':
          description: Successful Response
      description: Download the canonical YAML contract for the authenticated WebUI REST API.
      security:
      - WebUISession: []
      x-ktoolbox-mcp:
        enabled: false
        name: download_openapi
        scope: null
        safety: null
        openWorld: false
  /api/v1/mcp/status:
    get:
      tags:
      - mcp
      summary: Get MCP status
      operationId: mcp_status
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MCPStatusResponse'
      description: Return the in-process MCP endpoint, transport, and curated tool count.
      security:
      - WebUISession: []
      x-ktoolbox-mcp:
        enabled: false
        name: mcp_status
        scope: null
        safety: null
        openWorld: false
  /api/v1/mcp/tools:
    get:
      tags:
      - mcp
      summary: List MCP tools
      operationId: list_mcp_tools
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/MCPToolResponse'
                type: array
                title: Response List Mcp Tools
      description: Return the curated MCP tool catalog shown in the WebUI.
      security:
      - WebUISession: []
      x-ktoolbox-mcp:
        enabled: false
        name: list_mcp_tools
        scope: null
        safety: null
        openWorld: false
  /api/v1/mcp/tokens:
    get:
      tags:
      - mcp
      summary: List MCP tokens
      operationId: list_mcp_tokens
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/MCPTokenResponse'
                type: array
                title: Response List Mcp Tokens
      description: List MCP token metadata without returning token secrets.
      security:
      - WebUISession: []
      x-ktoolbox-mcp:
        enabled: false
        name: list_mcp_tokens
        scope: null
        safety: null
        openWorld: false
    post:
      tags:
      - mcp
      summary: Create an MCP token
      operationId: create_mcp_token
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MCPTokenCreateRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MCPTokenCreatedResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      description: Confirm the WebUI account password and create a one-time MCP bearer token.
      security:
      - WebUISession: []
      x-ktoolbox-mcp:
        enabled: false
        name: create_mcp_token
        scope: null
        safety: null
        openWorld: false
      parameters:
      - name: X-CSRF-Token
        in: header
        required: true
        description: CSRF token returned by the current browser session.
        schema:
          type: string
          minLength: 1
  /api/v1/mcp/tokens/{token_id}/revoke:
    post:
      tags:
      - mcp
      summary: Revoke an MCP token
      operationId: revoke_mcp_token
      parameters:
      - name: token_id
        in: path
        required: true
        schema:
          type: string
          title: Token Id
      - name: X-CSRF-Token
        in: header
        required: true
        description: CSRF token returned by the current browser session.
        schema:
          type: string
          minLength: 1
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MCPTokenResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      description: Immediately revoke an MCP bearer token owned by the current WebUI account.
      security:
      - WebUISession: []
      x-ktoolbox-mcp:
        enabled: false
        name: revoke_mcp_token
        scope: null
        safety: null
        openWorld: false
  /api/v1/mcp/files:
    get:
      tags:
      - project
      summary: Browse project files
      operationId: browse_project_files
      parameters:
      - name: mode
        in: query
        required: false
        schema:
          enum:
          - directory
          - file
          type: string
          default: directory
          title: Mode
      - name: path
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Path
      - name: search
        in: query
        required: false
        schema:
          type: string
          maxLength: 256
          default: ''
          title: Search
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 100
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FilesystemBrowseResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      description: Browse visible files and directories below the active project root without following paths outside it.
      security:
      - WebUISession: []
      x-ktoolbox-mcp:
        enabled: true
        name: browse_project_files
        scope: mcp:read
        safety: read
        openWorld: false
  /api/v1/mcp/config/{name}:
    patch:
      tags:
      - configuration
      summary: Patch safe configuration fields
      operationId: patch_mcp_configuration
      parameters:
      - name: name
        in: path
        required: true
        schema:
          enum:
          - dotenv
          - production
          type: string
          title: Name
      - name: locale
        in: query
        required: false
        schema:
          enum:
          - zh-CN
          - zh-Hant
          - en
          - ja
          - ko
          - fr
          - ru
          type: string
          default: en
          title: Locale
      - name: X-CSRF-Token
        in: header
        required: true
        description: CSRF token returned by the current browser session.
        schema:
          type: string
          minLength: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MCPConfigurationPatchRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfigSchemaResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      description: Update validated non-secret environment configuration fields by their documented field paths.
      security:
      - WebUISession: []
      x-ktoolbox-mcp:
        enabled: true
        name: patch_mcp_configuration
        scope: mcp:write
        safety: write
        openWorld: false
  /api/v1/mcp/tasks/{task_id}:
    delete:
      tags:
      - tasks
      summary: Delete a task record
      operationId: delete_task_record
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          title: Task Id
      - name: X-CSRF-Token
        in: header
        required: true
        description: CSRF token returned by the current browser session.
        schema:
          type: string
          minLength: 1
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskCleanupPreview'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      description: Delete an inactive task record and its logs while always preserving downloaded output files.
      security:
      - WebUISession: []
      x-ktoolbox-mcp:
        enabled: true
        name: delete_task_record
        scope: mcp:write
        safety: destructive
        openWorld: false
  /api/v1/mcp/pawchive/posts/{service}/{creator_id}/{post_id}:
    get:
      tags:
      - pawchive
      summary: Get a bounded Pawchive work
      operationId: get_mcp_work
      parameters:
      - name: service
        in: path
        required: true
        schema:
          type: string
          title: Service
      - name: creator_id
        in: path
        required: true
        schema:
          type: string
          title: Creator Id
      - name: post_id
        in: path
        required: true
        schema:
          type: string
          title: Post Id
      - name: revision_id
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Revision Id
      - name: include_content
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Include Content
      - name: content_limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 20000
          minimum: 1
          default: 12000
          title: Content Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                anyOf:
                - $ref: '#/components/schemas/MCPPostResponse'
                - $ref: '#/components/schemas/MCPRevisionResponse'
                title: Response Get Mcp Work
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      description: Return one Pawchive work with content omitted by default and explicitly capped when requested.
      security:
      - WebUISession: []
      x-ktoolbox-mcp:
        enabled: true
        name: get_work
        scope: mcp:read
        safety: read
        openWorld: true
...
  /api/v1/project:
    get:
      summary: Get project summary
      operationId: project
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectSummaryResponse'
      description: Return the active project root, configuration files, and KToolBox version.
      tags:
      - project
      security:
      - WebUISession: []
      x-ktoolbox-mcp:
        enabled: true
        name: get_project_summary
        scope: mcp:read
        safety: read
        openWorld: false
...
components:
  schemas:
...
    ConfigSchemaResponse:
      properties:
        locale:
          type: string
          enum:
          - zh-CN
          - zh-Hant
          - en
          - ja
          - ko
          - fr
          - ru
          title: Locale
        sections:
          additionalProperties:
            type: string
          type: object
          title: Sections
        fields:
          items:
            $ref: '#/components/schemas/ConfigFieldResponse'
          type: array
          title: Fields
      type: object
      required:
      - locale
      - sections
      - fields
      title: ConfigSchemaResponse
...
    FilesystemBrowseResponse:
      properties:
        scope:
          type: string
          enum:
          - project
          - host
          title: Scope
        mode:
          type: string
          enum:
          - directory
          - file
          title: Mode
        path:
          type: string
          title: Path
        project_relative_path:
          anyOf:
          - type: string
          - type: 'null'
          title: Project Relative Path
        parent:
          anyOf:
          - type: string
          - type: 'null'
          title: Parent
        separator:
          type: string
          title: Separator
        breadcrumbs:
          items:
            $ref: '#/components/schemas/FilesystemBreadcrumbResponse'
          type: array
          title: Breadcrumbs
        locations:
          items:
            $ref: '#/components/schemas/FilesystemLocationResponse'
          type: array
          title: Locations
        entries:
          items:
            $ref: '#/components/schemas/FilesystemEntryResponse'
          type: array
          title: Entries
        suggested_name:
          anyOf:
          - type: string
          - type: 'null'
          title: Suggested Name
        offset:
          type: integer
          title: Offset
        limit:
          type: integer
          title: Limit
        has_more:
          type: boolean
          title: Has More
      type: object
      required:
      - scope
      - mode
      - path
      - separator
      - breadcrumbs
      - locations
      - entries
      - offset
      - limit
      - has_more
      title: FilesystemBrowseResponse
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    MCPConfigurationPatchRequest:
      properties:
        values:
          additionalProperties:
            anyOf:
            - type: string
            - type: integer
            - type: number
            - type: boolean
            - items:
                type: string
              type: array
            - type: 'null'
          type: object
          minProperties: 1
          title: Values
      type: object
      required:
      - values
      title: MCPConfigurationPatchRequest
    MCPPostResponse:
      properties:
        id:
          type: string
          minLength: 1
          title: Id
        user:
          type: string
          minLength: 1
          title: User
        service:
          type: string
          minLength: 1
          title: Service
        title:
          anyOf:
          - type: string
          - type: 'null'
          title: Title
        content:
          anyOf:
          - type: string
          - type: 'null'
          title: Content
        substring:
          anyOf:
          - type: string
          - type: 'null'
          title: Substring
        embed:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Embed
        shared_file:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Shared File
        added:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Added
        published:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Published
        edited:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Edited
        file:
          anyOf:
          - $ref: '#/components/schemas/FileReference'
          - type: 'null'
        attachments:
          anyOf:
          - items:
              $ref: '#/components/schemas/FileReference'
            type: array
          - type: 'null'
          title: Attachments
        poll:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Poll
        captions:
          anyOf:
          - items: {}
            type: array
          - type: 'null'
          title: Captions
        tags:
          anyOf:
          - type: string
          - items: {}
            type: array
          - type: 'null'
          title: Tags
        origin:
          anyOf:
          - type: string
          - type: 'null'
          title: Origin
        preview_state:
          anyOf:
          - type: string
          - type: 'null'
          title: Preview State
        has_full:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Has Full
        preview_attempts:
          anyOf:
          - type: integer
          - type: 'null'
          title: Preview Attempts
        detail_fetched:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Detail Fetched
        import_size_cap_gb:
          anyOf:
          - type: number
          - type: 'null'
          title: Import Size Cap Gb
        next:
          anyOf:
          - type: string
          - type: 'null'
          title: Next
        prev:
          anyOf:
          - type: string
          - type: 'null'
          title: Prev
        effective_published:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Effective Published
        published_service_timezone:
          type: string
          title: Published Service Timezone
        published_target_timezone:
          type: string
          title: Published Target Timezone
      additionalProperties: true
      type: object
      required:
      - id
      - user
      - service
      - published_service_timezone
      - published_target_timezone
      title: MCPPostResponse
...
    MCPStatusResponse:
      properties:
        running:
          type: boolean
          title: Running
          default: true
        endpoint_path:
          type: string
          title: Endpoint Path
          default: /mcp
        openapi_path:
          type: string
          title: Openapi Path
          default: /api/v1/openapi.yaml
        transport:
          type: string
          const: streamable-http
          title: Transport
          default: streamable-http
        tool_count:
          type: integer
          title: Tool Count
      type: object
      required:
      - tool_count
      title: MCPStatusResponse
    MCPTokenCreateRequest:
      properties:
        name:
          type: string
          maxLength: 100
          minLength: 1
          title: Name
        password:
          type: string
          maxLength: 4096
          minLength: 1
          title: Password
        permission:
          type: string
          enum:
          - read
          - manage
          title: Permission
        expires_in_days:
          anyOf:
          - type: integer
            enum:
            - 7
            - 30
            - 90
            - 365
          - type: 'null'
          title: Expires In Days
      type: object
      required:
      - name
      - password
      - permission
      title: MCPTokenCreateRequest
    MCPTokenCreatedResponse:
      properties:
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
        username:
          type: string
          title: Username
        permission:
          type: string
          enum:
          - read
          - manage
          title: Permission
        scopes:
          items:
            type: string
          type: array
          title: Scopes
        created_at:
          type: string
          format: date-time
          title: Created At
        expires_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Expires At
        last_used_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Used At
        revoked_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Revoked At
        active:
          type: boolean
          title: Active
        token:
          type: string
          title: Token
      type: object
      required:
      - id
      - name
      - username
      - permission
      - scopes
      - created_at
      - expires_at
      - last_used_at
      - revoked_at
      - active
      - token
      title: MCPTokenCreatedResponse
    MCPTokenResponse:
      properties:
        id:
          type: string
          title: Id
        name:
          type: string
          title: Name
        username:
          type: string
          title: Username
        permission:
          type: string
          enum:
          - read
          - manage
          title: Permission
        scopes:
          items:
            type: string
          type: array
          title: Scopes
        created_at:
          type: string
          format: date-time
          title: Created At
        expires_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Expires At
        last_used_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Used At
        revoked_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Revoked At
        active:
          type: boolean
          title: Active
      type: object
      required:
      - id
      - name
      - username
      - permission
      - scopes
      - created_at
      - expires_at
      - last_used_at
      - revoked_at
      - active
      title: MCPTokenResponse
    MCPToolResponse:
      properties:
        name:
          type: string
          title: Name
        operation_id:
          type: string
          title: Operation Id
        description:
          type: string
          title: Description
        category:
          type: string
          title: Category
        scope:
          type: string
          enum:
          - mcp:read
          - mcp:write
          title: Scope
        safety:
          type: string
          enum:
          - read
          - write
          - destructive
          title: Safety
        open_world:
          type: boolean
          title: Open World
      type: object
      required:
      - name
      - operation_id
      - description
      - category
      - scope
      - safety
      - open_world
      title: MCPToolResponse
...
    TaskCleanupPreview:
      properties:
        task_id:
          type: string
          title: Task Id
        artifacts:
          items:
            $ref: '#/components/schemas/TaskArtifact'
          type: array
          title: Artifacts
        removable_files:
          type: integer
          title: Removable Files
        removable_bytes:
          type: integer
          title: Removable Bytes
      type: object
      required:
      - task_id
      - artifacts
      - removable_files
      - removable_bytes
      title: TaskCleanupPreview
...
  securitySchemes:
    WebUISession:
      type: apiKey
      in: cookie
      name: ktoolbox_session
      description: Opaque browser session created by the login operation.
...